:pencil: use correct c++

+++ main.cpp
@@ -3,6 +3,7 @@
 //
 
 #include <iostream>
+#include <FlexLexer.h>
 #include "parser.yy.hpp"
 
 
@@ -10,7 +11,7 @@ int main(int argc, char **argv)
 {
     try
     {
-        for (std::size_t index = 1; index<argc; ++index)
+        for (std::size_t index = 1; index < static_cast<std::size_t >(argc); ++index)
         {
             std::cout << "Reading : " << argv[index] << std::endl;
             freopen(argv[index], "r", stdin);
